Workspace prompt.

+++ products/org.csstudio.iter.css.product/src/main/java/org/csstudio/iter/css/product/ITERApplication.java
@@ -14,8 +14,10 @@ import java.net.URL;
 
 import org.csstudio.startup.application.Application;
 import org.eclipse.core.runtime.Platform;
+import org.eclipse.equinox.app.IApplication;
 import org.eclipse.equinox.app.IApplicationContext;
 import org.eclipse.osgi.service.datalocation.Location;
+import org.eclipse.swt.widgets.Display;
 
 /**
  * 
@@ -28,7 +30,11 @@ import org.eclipse.osgi.service.datalocation.Location;
 public class ITERApplication extends Application {
 
     @Override
-    public Object start(IApplicationContext context) throws Exception {    
+    protected Object promptForWorkspace(Display display, IApplicationContext context) throws Exception {
+        Object obj = super.promptForWorkspace(display, context);
+        if (obj == IApplication.EXIT_OK) {
+            return obj;
+        }
         final String args[] = (String[]) context.getArguments().get("application.args");
         for (String arg : args) {
             if (arg.equals("-cleanWorkbench")) { 
@@ -44,28 +50,35 @@ public class ITERApplication extends Application {
         }
         
         Location loc = Platform.getInstanceLocation();
-        URL url = loc.getURL();
-        File file = new File(url.getFile());
-        file = new File(file, ".metadata");
-        file = new File(file, ".plugins");
-        file = new File(file, "org.eclipse.e4.workbench");
-        file = new File(file, "workbench.xmi");
-        if (file.exists()) {
-            boolean delete = false;
-            try (BufferedReader br = new BufferedReader(new FileReader(file))) {
-                while(br.ready()) {
-                    String line = br.readLine();
-                    if (line != null && line.contains("org.csstudio.alarm.beast.ui.alarmtable")) {
-                        delete = true;
-                        break;
+        if (loc.isSet()) {
+            URL url = loc.getURL();
+            File file = new File(url.getFile());
+            file = new File(file, ".metadata");
+            file = new File(file, ".plugins");
+            file = new File(file, "org.eclipse.e4.workbench");
+            file = new File(file, "workbench.xmi");
+            if (file.exists()) {
+                boolean delete = false;
+                try (BufferedReader br = new BufferedReader(new FileReader(file))) {
+                    while(br.ready()) {
+                        String line = br.readLine();
+                        if (line != null && line.contains("org.csstudio.alarm.beast.ui.alarmtable")) {
+                            delete = true;
+                            break;
+                        }
                     }
                 }
+                if (delete) {
+                    file.delete();
+                }
             }
-            if (delete) {
-                file.delete();
-            }
-            
         }
+        
+        return obj;
+    }
+    
+    @Override
+    public Object start(IApplicationContext context) throws Exception {    
         //org.csstudio.alarm.beast.ui.alarmtable
         Object o = super.start(context);
         //Bugfix/workaround for org.apache.felix.gogo.shell.Activator,
